Add stored date display when queried
This adds support for the stored date to be displayed when querying a
stored vehicle.

Currently there is no way to convert the older vehicles to the new
format so they will display as "Store date invalid" till they are
restored.

Thanks to @dietnx for the pestering to get this done.