Add support for removing weapons on vehicle spawn
This is quite a requested feature since players are abusing the garage
to get free rearms.

Basically what this does is when the player stores a vehicle, the server
(on server startup) will generate a random key and store that in the
database.

When the vehicle is spawned, it will check that key against the current
server key, if they match, it means the player stored that vehicle this
restart and the ammo needs to be stripped, if they don't match this
vehicle was obviously stored on a previous restart.

Thanks to @icomrade/@Dayz_Epoch for the random key code and to @icomrade
again for the heads up on the turret code on the biki.